+++ tests/functional/data/dockerfiles/valid/sleep/sleep.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
 echo before sleep
 # heartbeat is every second so 2 seconds is enough to get at least one heartbeat
-sleep 2
+sleep ${SLEEP:-2}
 echo after sleep